requirements.txt
groq
sentence-transformers
streamlit
pymongo
bcrypt
pinecone-client==3.0.0
PyPDF2
